# docs/swagger.yaml
definitions:
  models.AdminInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  models.ConfigurationInput:
    properties:
      artifact:
        additionalProperties: true
        type: object
      description:
        type: string
      input_protein:
        type: string
      is_notification_on:
        type: boolean
      lab_result:
        $ref: '#/definitions/models.LabResult'
      meta:
        items:
          type: string
        type: array
      name:
        type: string
      options:
        additionalProperties: true
        type: object
      ref_job_id:
        type: string
      run_type:
        type: string
      stage_id:
        maximum: 3
        minimum: 0
        type: integer
      state:
        type: string
      user_id:
        type: string
    required:
    - input_protein
    - name
    - options
    - ref_job_id
    - run_type
    - user_id
    type: object
  models.CopyQueryResultInput:
    properties:
      complete_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      input_protein:
        type: string
      job_id:
        type: string
      result:
        items:
          $ref: '#/definitions/models.CopyResultInput'
        type: array
      run_id:
        type: integer
      state:
        type: string
    type: object
  models.CopyResultInput:
    properties:
      acc_len:
        type: integer
      accession:
        type: string
      description:
        type: string
      e_values:
        type: number
      hsp_query_from:
        type: number
      hsp_query_to:
        type: number
      id:
        type: string
      is_selected:
        type: boolean
      max_score:
        type: number
      organisms:
        type: string
      percent_identity:
        type: number
      query_cover:
        type: number
      score:
        type: number
      sequences:
        type: string
    type: object
  models.CreateJobInput:
    properties:
      job:
        $ref: '#/definitions/models.JobInput'
      query_result:
        $ref: '#/definitions/models.CopyQueryResultInput'
    required:
    - job
    type: object
...
  models.HttpResponseError:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.HttpResponseOK:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  models.JobInput:
    properties:
      artifact:
        additionalProperties: true
        type: object
      description:
        type: string
      input_protein:
        type: string
      is_notification_on:
        type: boolean
      lab_result:
        $ref: '#/definitions/models.LabResult'
      meta:
        items:
          type: string
        type: array
      name:
        type: string
      options:
        additionalProperties: true
        type: object
      ref_job_id:
        type: string
      run_type:
        type: string
      stage_id:
        maximum: 3
        minimum: 0
        type: integer
    required:
    - input_protein
    - name
    - options
    - run_type
    type: object
  models.LabResult:
    properties:
      scores:
        items:
          type: number
        type: array
      sequences:
        items:
          type: string
        type: array
      total:
        minimum: 0
        type: integer
    required:
    - scores
    - sequences
    type: object
...
  models.SignInAdminInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact:
    email: proteng.plus@gmail.com
    name: Proteng API Support
  title: Proteng API
  version: "1.0"
paths:
...
  /proteng-conductor/jobs:
    get:
      description: Retrieve a list of all jobs
      parameters:
      - description: Filter state of the jobs
        in: query
        name: state
        type: string
      - description: Filter name of the job
        in: query
        name: name
        type: string
      - description: Sort by
        in: query
        name: sort
        type: string
      - description: Sort direction (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get all jobs
      tags:
      - Jobs
    post:
      consumes:
      - application/json
      description: Create a new job with the provided details
      parameters:
      - description: Create object
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.CreateJobInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new job
      tags:
      - Jobs
  /proteng-conductor/jobs/{id}:
    delete:
      description: Delete a job by its ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a job by ID
      tags:
      - Jobs
    get:
      description: Retrieve details of a job by its ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a job by ID
      tags:
      - Jobs
    put:
      consumes:
      - application/json
      description: Update an existing job with the provided details
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Job object
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.JobInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update an existing job
      tags:
      - Jobs
  /proteng-conductor/jobs/{id}/run:
    post:
      description: Start a job by its ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Start a job by ID
      tags:
      - Jobs
  /proteng-conductor/jobs/configurations:
    get:
      description: Retrieve a list of all configurations
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get all configurations
      tags:
      - Configurations
    post:
      consumes:
      - application/json
      description: Create a new configuration with the provided details
      parameters:
      - description: Configuration object
        in: body
        name: configuration
        required: true
        schema:
          $ref: '#/definitions/models.ConfigurationInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Save a new configuration
      tags:
      - Configurations
  /proteng-conductor/jobs/dashboard:
    get:
      description: Retrieve information about the jobs to show in the dashboard page
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get job dashboard
      tags:
      - Jobs
...
  /proteng-user-mgmt/admins:
    get:
      description: Retrieve a list of all admins
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all admins
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates a new admin with the provided details
      parameters:
      - description: Admin object
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Create a new admin
      tags:
      - Admin
  /proteng-user-mgmt/admins/{id}:
    delete:
      description: Delete an admin by providing its ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete an admin
      tags:
      - Admin
    get:
      description: Retrieve an admin by providing its ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve an admin by ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update an existing admin with the provided details
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: Admin object
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update an existing admin
      tags:
      - Admin
...
  /proteng-user-mgmt/auth/login/admin:
    post:
      consumes:
      - application/json
      description: Signs in an admin by validating the provided credentials
      parameters:
      - description: Admin credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.SignInAdminInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Sign in an admin
      tags:
      - Authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"